Use the revamped help system scheduled for v5 release w/ Lita v4.7.0
Lita v5 is still a work in progress, but we want to make use of its new
help system. Forking and replacing the built in help handler with the
new code as seen in lita-default-handlers is the most painless solution,
and is a temporary fix until Lita v5 is released (scheduled for this
month).