Trigger :message_dispatched in dispatch_to_route instead of dispatch
This addresses DEV-878 (lita-metrics not logging valid commands
requiring confirmation). I'm going to open a PR on litaio/lita with the
same change, but if that gets merged, it will be merged into master.
Since we're on our own fork of Lita v4.7.0, those changes won't be
pulled in to OfficerURL automatically.

See: https://github.com/jimmycuadra/lita-confirmation/pull/9