[DEV-921] Modify regex used by the Message class to detect a command
See https://github.com/litaio/lita/pull/187

The previous regex allowed each line beginning with Lita's mention name
or alias in a multi-line message to be interpreted as a command, which
is problematic for commands deliberately enclosed in code blockquotes
like "\n!foo\n". With this change, Lita will check for the
presence of a command once per message instead of once per line.